pyproject.toml
[project]
name = "heidi-cli"
version = "0.1.0"
description = "Heidi CLI - orchestrate Copilot SDK + Jules + OpenCode agent loops"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
  "github-copilot-sdk>=0.1.23",
  "keyring>=24.0.0",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  "requests>=2.31.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.3.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
heidi = "heidi_cli.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]

